feature: add docker container
