+++ docker_requirements.txt
@@ -1,0 +1,12 @@
+ipython==8.14.0
+kaleido==0.2.1
+numpy==1.24.0
+pandas==2.0.2
+pytest==7.4.3
+plotly==5.15.0
+scikit_learn==1.3.2
+scipy==1.8.0
+tqdm==4.66.1
+cloudpickle==3.0.0
+pivottablejs==0.9.0
+streamlit==1.28.0
